ready-set-realtor/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
celery==5.3.6
redis==5.0.1
supabase==2.0.3
langchain==0.0.350
openai==1.3.7
python-dotenv==1.0.0
pytest==7.4.3
httpx>=0.24.0,<0.25.0 